P2 batch 4: type corrections (TC-01/02/06), trait migrations (TM-05/06), integrations (INT-11/13/14/15/16/20), naming/infra (NF-01-06, IF-01/03)
Type corrections:
- TC-01: Document Decay::Ttl relative-vs-absolute design decision
- TC-02: Replace `tainted: bool` with typed `Taint` enum (9 variants, #[non_exhaustive])
  with backward-compat TaintInfo bridge and bidirectional From impls
- TC-06: Consolidate PadVector to canonical f64 definition in roko-primitives,
  both roko-core and roko-runtime re-export from single source

Trait migrations (additive, non-breaking):
- TM-05: Composer::compose_datums(&[Datum<'_>]) with default impl filtering to engrams
- TM-06: Policy::decide_with_pulses(&[Engram], &[Pulse]) -> PolicyOutputs with default impl

Integrations (6 new wiring paths):
- INT-11: Plan completion events -> daimon affect appraisal (somatic outcomes)
- INT-13: Conductor health degradation -> pheromone field (Anomaly/Threat deposits)
- INT-14: Safety constraints -> system prompt anti-patterns (layer 7 injection)
- INT-15: Neuro failure/stability knowledge -> adaptive gate thresholds (CUSUM tuning)
- INT-16: Code-intel hints -> verification gate (rung 3 symbol enrichment)
- INT-20: Agent lifecycle transitions -> neuro knowledge store (anti-knowledge for failures)

Infrastructure:
- 12 property tests for roko-conductor (CircuitBreaker + YerkesDodson)
- MSRV CI workflow checking Rust 1.85 compatibility
- Fix f32/f64 mismatches in theta_consumer after PadVector consolidation

All builds clean (0 errors, 0 warnings). All tests pass.

+++ crates/roko-runtime/src/heartbeat_attention.rs
@@ -93,11 +93,11 @@ impl ContextCategory {
 
     fn affect_multiplier(self, pad: &PadVector) -> f64 {
         match self {
-            Self::Safety => 1.0 + f64::from(pad.arousal).abs() * 0.5,
+            Self::Safety => 1.0 + pad.arousal.abs() * 0.5,
             Self::Exploration | Self::ResearchArtifacts => {
-                1.0 + (1.0 - f64::from(pad.dominance)) * 0.3
+                1.0 + (1.0 - pad.dominance) * 0.3
             }
-            Self::IterationMemory => 1.0 + f64::from(-pad.pleasure).max(0.0) * 0.4,
+            Self::IterationMemory => 1.0 + (-pad.pleasure).max(0.0) * 0.4,
             _ => 1.0,
         }
     }
